Added a Navigation Bar
It is a semi functional nav bar. Only linked it to 2 of my pages and Diman's quiz page for now. I'll link everything else once we get this thing going. I have added a <! navigation bar> section to quiz, query pages as well. do not touch it. start building the website below it.